Move pyramid and celery app fixtures to conftest
Move scoped_pyramid_app and celery_app from test_subscribers.py to
conftest.py so they are available for other tests.
